windows 10, wheel finding, and more
Start device.write commands with a 0x00 instead of relying on library
code to do that for us. This fixes Windows 10. Mac and Linux seem happy
with it too.

Fix for odd logic in forceOff()

More accurate targeting on each OS to find exactly which instance of a
G29 we want to connect to.